Fixed QA comments
