+++ libs/dh/globalization/assets-localization/src/assets/i18n/da.json
@@ -1204,6 +1204,8 @@
       "balanceResponsible": "Balanceansvarlig",
       "error": "Der opstod desværre en fejl, under anmodingen om beregning",
       "startDateCannotBeAfterEndDate": "Startdato kan ikke være efter slutdato",
+      "startDateCannotBeOlderThan3Years": "Startdato kan ikke være ældre end 3 år",
+      "startAndEndDateCannotBeInTheFuture": "Startdato og slutdato kan ikke være i fremtiden",
       "success": "Din anmoding er startet.",
       "topBarTitle": "Anmod om beregning",
       "processType": "{{shared.processType}}",
