+++ libs/dh/globalization/assets-localization/src/assets/i18n/en.json
@@ -1211,6 +1211,9 @@
       "energySupplier": "Energy supplier",
       "balanceResponsible": "Balance responsible",
       "error": "Unfortunately, an error occurred requesting the calculation",
+      "startDateCannotBeAfterEndDate": "Start date cannot be after end date",
+      "startDateCannotBeOlderThan3Years": "Start date cannot be older than 3 years",
+      "startAndEndDateCannotBeInTheFuture": "Start and end date cannot be in the future",
       "success": "The calculation has been requested",
       "topBarTitle": "Request calculation",
       "processType": "{{shared.processType}}",
